requirements.txt
aws-cdk-lib==2.38.1
constructs>=10.0.0,<11.0.0
boto3
Flask==2.2.2
